move train module one level up in package structure
