+++ tests/test_rf_serial_synthetic.py
@@ -4,8 +4,8 @@ import shutil
 
 import pytest
 
+from specialcouscous.train import train_serial_on_synthetic_data
 from specialcouscous.utils import set_logger_config
-from specialcouscous.utils.train import train_serial_on_synthetic_data
 
 log = logging.getLogger("specialcouscous")  # Get logger instance.
